+++ src/fidesops/schemas/saas/shared_schemas.py
@@ -31,3 +31,15 @@ class SaaSRequestParams(BaseModel):
         """Using enum values"""
 
         use_enum_values = True
+
+
+class ConnectorParamRef(BaseModel):
+    """A reference to a value in the connector params (by name)"""
+
+    connector_param: Any
+
+
+class IdentityParamRef(BaseModel):
+    """A reference to the identity type in the filter Post Processor Config"""
+
+    identity: str