+++ src/fidesops/service/authentication/authentication_strategy_query_param.py
@@ -1,7 +1,6 @@
-from typing import Any, Dict
-
 from requests import PreparedRequest
 
+from fidesops.models.connectionconfig import ConnectionConfig
 from fidesops.schemas.saas.strategy_configuration import (
     QueryParamAuthenticationConfiguration,
     StrategyConfiguration,
@@ -26,11 +25,13 @@ class QueryParamAuthenticationStrategy(AuthenticationStrategy):
         self.value = configuration.value
 
     def add_authentication(
-        self, request: PreparedRequest, secrets: Dict[str, Any]
+        self, request: PreparedRequest, connection_config: ConnectionConfig
     ) -> PreparedRequest:
         """Add token to the request as a query param"""
         request.url = set_query_parameter(
-            request.url, self.name, assign_placeholders(self.value, secrets)
+            request.url,
+            self.name,
+            assign_placeholders(self.value, connection_config.secrets),
         )
         return request