+++ tests/api/v1/endpoints/test_saas_config_endpoints.py
@@ -1,16 +1,21 @@
 import json
 from typing import Optional
+from unittest import mock
+from unittest.mock import Mock
 
 import pytest
 from sqlalchemy.orm import Session
 from starlette.testclient import TestClient
 
 from fidesops.api.v1.scope_registry import (
+    CLIENT_READ,
+    CONNECTION_AUTHORIZE,
     SAAS_CONFIG_CREATE_OR_UPDATE,
     SAAS_CONFIG_DELETE,
     SAAS_CONFIG_READ,
 )
 from fidesops.api.v1.urn_registry import (
+    AUTHORIZE,
     SAAS_CONFIG,
     SAAS_CONFIG_VALIDATE,
     V1_URL_PREFIX,
@@ -408,3 +413,39 @@ class TestDeleteSaaSConfig:
             "before deleting this SaaS config. Must clear the secrets from this connection "
             "config before deleting the SaaS config."
         )
+
+
+class TestAuthorizeConnection:
+    @pytest.fixture
+    def authorize_url(self, oauth2_connection_config) -> str:
+        path = V1_URL_PREFIX + AUTHORIZE
+        path_params = {"connection_key": oauth2_connection_config.key}
+        return path.format(**path_params)
+
+    def test_client_not_authenticated(self, api_client: TestClient, authorize_url):
+        response = api_client.get(authorize_url)
+        assert response.status_code == 401
+
+    def test_client_wrong_scope(
+        self, api_client: TestClient, authorize_url, generate_auth_header
+    ) -> None:
+        auth_header = generate_auth_header([CLIENT_READ])
+        response = api_client.get(authorize_url, headers=auth_header)
+        assert 403 == response.status_code
+
+    @mock.patch(
+        "fidesops.api.v1.endpoints.saas_config_endpoints.OAuth2AuthenticationStrategy.get_authorization_url"
+    )
+    def test_get_authorize_url(
+        self,
+        authorization_url_mock: Mock,
+        api_client: TestClient,
+        authorize_url,
+        generate_auth_header,
+    ):
+        authorization_url = "https://localhost/auth/authorize"
+        authorization_url_mock.return_value = authorization_url
+        auth_header = generate_auth_header([CONNECTION_AUTHORIZE])
+        response = api_client.get(authorize_url, headers=auth_header)
+        assert response.ok
+        assert response.text == f'"{authorization_url}"'