+++ tests/service/authentication/test_authentication_strategy_basic.py
@@ -2,6 +2,7 @@ import pytest
 from requests import PreparedRequest, Request
 
 from fidesops.common_exceptions import ValidationError as FidesopsValidationError
+from fidesops.models.connectionconfig import ConnectionConfig
 from fidesops.service.authentication.authentication_strategy_factory import get_strategy
 from fidesops.util.cryptographic_util import bytes_to_b64_str
 
@@ -15,7 +16,7 @@ def test_basic_auth_with_username_and_password():
 
     authenticated_request = get_strategy(
         "basic", {"username": "<username>", "password": "<password>"}
-    ).add_authentication(req, secrets)
+    ).add_authentication(req, ConnectionConfig(secrets=secrets))
     assert (
         authenticated_request.headers["Authorization"]
         == f"Basic {bytes_to_b64_str(bytes(f'{username}:{password}', 'UTF-8'))}"
@@ -30,7 +31,7 @@ def test_basic_auth_with_username_only():
 
     authenticated_request = get_strategy(
         "basic", {"username": "<username>"}
-    ).add_authentication(req, secrets)
+    ).add_authentication(req, ConnectionConfig(secrets=secrets))
     # The requests library still calls str(password) even if the password is None
     assert (
         authenticated_request.headers["Authorization"]
@@ -42,4 +43,4 @@ def test_basic_auth_with_no_credentials():
     req: PreparedRequest = Request(method="POST", url="https://localhost").prepare()
 
     with pytest.raises(FidesopsValidationError):
-        get_strategy("basic", {}).add_authentication(req, {})
+        get_strategy("basic", {}).add_authentication(req, ConnectionConfig(secrets={}))